requirements.txt
streamlit
python-dotenv
google-generativeai
langchain # For RAG
langchain-google-genai # For Gemini embeddings & chat models
chromadb # For vector store (will use SQLite backend)
pypdf # For PDF parsing
python-docx # Optional, for Word docs
SQLAlchemy # For database ORM
aiosqlite # Async driver for SQLite
PyJWT # For JWT handling (if we do simple manual tokens)
passlib[bcrypt] # For password hashing